:recycle: Separate views into subdirectories
